requirements.txt
APIFlask==2.3.1
apispec==6.8.0
Authlib==1.3.2
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
Flask-OAuthlib==0.9.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
msgspec==0.18.6
oauthlib==2.1.0
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pylimbo==0.0.10
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rsa==4.9
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
webargs==8.6.0
Werkzeug==3.1.3
zipp==3.21.0
